remove coppery object composer, migrate other coppery components back to nem2 sdk, delete coppery library dependency, use native JsonSerializer, update namespaces accordingly, +other small fixes. reserve Coppery namespace on httpclients

+++ nem2-sdk/Infrastructure/Responses/Pagination.cs
@@ -3,6 +3,6 @@
     public class Pagination
     {
         public int PageSize { get; set; }
-        public int PageNumber { get; set; }
+        public ulong PageNumber { get; set; }
     }
 }
